sched/nohz: fix overflow error in scheduler_tick_max_deferment()
The conversion of the max deferment from usecs to nsecs can easily
overflow on platforms where a long is 32-bits.  To fix, cast the usecs
value to u64 before multiplying by NSECS_PER_USEC.

This was discovered on 32-bit ARM platform when extending the max
deferment value.

Cc: Frederic Weisbecker <fweisbec@gmail.com>

+++ kernel/sched/core.c
@@ -2865,7 +2865,7 @@ u64 scheduler_tick_max_deferment(void)
 	if (time_before_eq(next, now))
 		return 0;
 
-	return jiffies_to_nsecs(next - now);
+	return (u64)jiffies_to_nsecs(next - now);
 }
 
 static __init int sched_nohz_full_init_debug(void)
